Implement driver.QueryerContext interface (#4)
As per the [specifications](https://pkg.go.dev/database/sql/driver#QueryerContext),
`QueryerContext` is an optional interface.
When it's not implemented, the driver falls back to the `Queryer` interface
that is also optional, and eventually defaults to:
  - preparing a query
  - executing the statement
  - closing the statement

This means that the context passed to `QueryContext` gets completely ignored
and therefore timeouts aren't honoured.

This commit provides an implementation of the `QueryerContext` interface
that honours the context. When the context expires or gets cancelled,
the statement gets cancelled and closed, and upon completion,
an error is returned.